# Conflicts:
#	src/xcode/ENA/ENA.xcodeproj/project.pbxproj

+++ src/xcode/ENA/ENAUITests/Home/ENAUITestsHome.swift
@@ -140,7 +140,7 @@ class ENAUITests_01_Home: XCTestCase {
 		
 		// find an element with localized text "Keine Risiko-Begegnungen"
 		let lowRiskTitle = String(format: AccessibilityLabels.localized(AppStrings.Home.riskCardLowNumberContactsItemTitle), numberOfDaysWithLowRisk)
-		XCTAssert(app.otherElements[lowRiskTitle].waitForExistence(timeout: .short))
+		XCTAssert(app.otherElements[lowRiskTitle].waitForExistence(timeout: .long))
 		
 		XCTAssert(app.buttons[AccessibilityIdentifiers.Home.rightBarButtonDescription].waitForExistence(timeout: .short))
 		snapshot("homescreenrisk_level_\(riskLevel)_risk_no_days_\(String(format: "%04d", (screenshotCounter.inc() )))")
